refactor: extract option system to minarch_options.c/h
Moves Option_getValueIndex, Option_setValue, and all OptionList_* functions
from minarch.c into a new minarch_options.c/h module. Promotes DEVICE_WIDTH,
DEVICE_HEIGHT, and Special_updatedDMGPalette to non-static with extern
declarations in minarch_internal.h to satisfy cross-module references.
